Extract site settings section navigation
