+++ assets/press-system.json
@@ -1,13 +1,13 @@
 {
   "schemaVersion": 1,
   "type": "press-system",
-  "version": "3.4.70",
-  "tag": "v3.4.70",
+  "version": "3.4.71",
+  "tag": "v3.4.71",
   "upgradeFrom": {
     "ranges": [
-      ">=3.4.69 <3.4.70"
+      ">=3.4.70 <3.4.71"
     ],
     "allowUnknownSource": false,
-    "message": "Update to v3.4.69 first."
+    "message": "Update to v3.4.70 first."
   }
 }
